refactor(omo): split omo_identifier.py into catalog + matching + prompt (Related to #1886)
590L monolith → 4 focused modules (TDD-first, 18 tests green):
- omo_lesson_catalog.py (225L) — LessonCandidate + cache loading + story-id resolution + hint path
- omo_title_matching.py (114L) — normalize_title + fuzzy_match_title (stdlib difflib, no Vertex AI)
- omo_identifier_prompt.py (63L) — _build_identification_prompt (pure string, no I/O)
- omo_identifier.py (297L) — thin async orchestrator (AI call + circuit breaker)

Public API unchanged: LessonCandidate, identify_lesson_from_image, identify_lesson_from_hint
all still importable from omo_identifier. Backward-compat: identify_lesson_from_hint kept in
omo_identifier so existing tests patching omo_identifier.* continue to work.
