+++ backend/app/services/omo_identifier.py
@@ -13,19 +13,28 @@ llm-endpoint-hardening checklist:
 - Fail-closed: returns empty candidates on error, never auto-passes ✅
 - Reasoning field: each candidate has reasoning string ✅
 - Circuit breaker: 3 consecutive errors → raise RuntimeError → 503 ✅
+
+Refactored (#1886): catalog/cache → omo_lesson_catalog.py,
+title normalisation + fuzzy match → omo_title_matching.py,
+prompt builder → omo_identifier_prompt.py.
 """
 
 import base64
-import difflib
 import json
 import logging
-from dataclasses import dataclass, field
-from pathlib import Path
-
-import yaml
 
 from .ai_base import _repair_json, GeminiContentFilterError
 from .llm_models import get_model_for_task
+from .omo_lesson_catalog import (
+    LessonCandidate,
+    _LESSON_CACHE,
+    _CURRICULUM_TITLE_CACHE,
+    _resolve_story_id,
+    _load_omo_lessons,
+    _load_curriculum_titles,
+)
+from .omo_title_matching import _normalize_title, _fuzzy_match_title
+from .omo_identifier_prompt import _build_identification_prompt
 
 logger = logging.getLogger(__name__)
 
@@ -35,312 +44,6 @@ logger = logging.getLogger(__name__)
 # Max consecutive AI errors before raising RuntimeError (circuit breaker)
 _MAX_CONSECUTIVE_ERRORS = 3
 _consecutive_errors = 0
-
-# Lesson metadata cache (loaded once at import)
-_LESSON_CACHE: list[dict] = []
-
-# Full curriculum title cache: {lesson_code -> {title, grade_code}} for all 158 lessons
-# Used by fuzzy-match fallback (stdlib difflib, no extra deps)
-_CURRICULUM_TITLE_CACHE: dict[str, dict] = {}
-
-
-@dataclass
-class LessonCandidate:
-    lesson_id: int
-    grade_code: str
-    title: str
-    confidence: float
-    reasoning: str = field(default="")
-    # #1775: canonical Story.id resolved at service boundary — None when grade_code
-    # is not in lesson_loader (e.g. fallback content-only lessons without a DB row).
-    story_id: int | None = field(default=None)
-
-
-def _resolve_story_id(grade_code: str) -> int | None:
-    """Resolve canonical Story.id from grade_code via lesson_loader.
-
-    Called once per candidate at identifier output construction so callers
-    receive story_id without needing to do a second lookup (#1775).
-    Returns None when grade_code has no matching DB-loaded lesson.
-    """
-    try:
-        from .lesson_loader import get_lesson_by_code
-        story = get_lesson_by_code(grade_code)
-        if story and story.get("id"):
-            return int(story["id"])
-    except Exception as exc:
-        logger.debug("_resolve_story_id: grade_code=%r lookup failed: %s", grade_code, exc)
-    return None
-
-
-def _load_omo_lessons() -> list[dict]:
-    """Load metadata for ALL lessons in the full curriculum catalog.
-
-    Source priority:
-    1. backend/data/curriculum/lessons/G*-L*.yml — 158 catalog entries (lesson_code, title, vocab, strategy_name)
-    2. Fallback: backend/data/lessons/L*.yml — 57 content lessons with story_text (for any lessons not in catalog)
-
-    Catalog entries provide title + vocab only (no story_text). Content lessons additionally provide story snippet.
-    """
-    global _LESSON_CACHE
-    if _LESSON_CACHE:
-        return _LESSON_CACHE
-
-    base_dir = Path(__file__).parent.parent.parent / "data"
-    catalog_dir = base_dir / "curriculum" / "lessons"
-    content_dir = base_dir / "lessons"
-
-    # Build content lookup by title (to enrich catalog entries with story_text)
-    content_by_title = {}
-    for yml_path in content_dir.glob("L*.yml"):
-        try:
-            with open(yml_path, "r", encoding="utf-8") as f:
-                data = yaml.safe_load(f) or {}
-            title = (data.get("title", "") or "").lstrip("-").strip()
-            if title:
-                content_by_title[title] = data
-        except Exception:
-            continue
-
-    result = []
-    # 1) Catalog scan (preferred — 158 lessons)
-    for yml_path in sorted(catalog_dir.glob("G*-L*.yml")):
-        try:
-            with open(yml_path, "r", encoding="utf-8") as f:
-                data = yaml.safe_load(f) or {}
-            title = (data.get("title", "") or "").strip()
-            if not title:
-                continue
-            lesson_code = data.get("lesson_code", yml_path.stem)
-            # Use lesson_code hash for lesson_id (stable int per lesson_code)
-            # Or parse from order field
-            lesson_id = data.get("display_order") or data.get("original_order") or 0
-            vocab_list = data.get("vocab") or []
-            # Enrich with story snippet if content YAML exists
-            story_snippet = ""
-            content_data = content_by_title.get(title)
-            if content_data:
-                story_snippet = (content_data.get("story_text", "") or "")[:120]
-            result.append({
-                "lesson_id": int(lesson_id) if lesson_id else 0,
-                "lesson_code": lesson_code,
-                "grade_code": lesson_code,
-                "title": title,
-                "story_snippet": story_snippet,
-                "vocabulary": [str(v) for v in vocab_list[:5]],
-            })
-        except Exception as exc:
-            logger.warning("Failed to load OMO catalog %s: %s", yml_path.name, exc)
-
-    # 2) Add any content-only lessons not in catalog (rare)
-    catalog_titles = {r["title"] for r in result}
-    for yml_path in content_dir.glob("L*.yml"):
-        try:
-            with open(yml_path, "r", encoding="utf-8") as f:
-                data = yaml.safe_load(f) or {}
-            title = (data.get("title", "") or "").lstrip("-").strip()
-            if not title or title in catalog_titles:
-                continue
-            stem = yml_path.stem
-            try:
-                lesson_id = int(stem[1:])
-            except ValueError:
-                continue
-            result.append({
-                "lesson_id": lesson_id,
-                "lesson_code": data.get("grade_code", f"L{lesson_id}"),
-                "grade_code": data.get("grade_code", f"L{lesson_id}"),
-                "title": title,
-                "story_snippet": (data.get("story_text", "") or "")[:120],
-                "vocabulary": [
-                    v.get("word", "") for v in (data.get("vocabulary") or [])[:5]
-                ],
-            })
-        except Exception as exc:
-            logger.warning("Failed to load fallback content lesson %s: %s", yml_path.name, exc)
-
-    _LESSON_CACHE = result
-    logger.info("Loaded %d OMO lesson entries (curriculum catalog + content fallback)", len(result))
-    return result
-
-
-def _load_curriculum_titles() -> dict[str, dict]:
-    """Load all curriculum lesson titles for fuzzy-match fallback.
-
-    Returns a dict of {lesson_code: {"title": str, "grade_code": str}} for all
-    lessons in backend/data/curriculum/lessons/*.yml.  Cached after first call.
-    Uses stdlib only (no extra deps).
-    """
-    global _CURRICULUM_TITLE_CACHE
-    if _CURRICULUM_TITLE_CACHE:
-        return _CURRICULUM_TITLE_CACHE
-
-    curriculum_dir = Path(__file__).parent.parent.parent / "data" / "curriculum" / "lessons"
-    if not curriculum_dir.exists():
-        logger.warning("Curriculum lessons dir not found: %s", curriculum_dir)
-        return {}
-
-    result: dict[str, dict] = {}
-    for yml_path in curriculum_dir.glob("*.yml"):
-        lesson_code = yml_path.stem  # e.g. "G5-L25"
-        try:
-            with open(yml_path, "r", encoding="utf-8") as f:
-                data = yaml.safe_load(f)
-            title = (data.get("title") or "").lstrip("-").strip()
-            grade_code = data.get("lesson_code") or lesson_code
-            if title:
-                result[lesson_code] = {"title": title, "grade_code": grade_code}
-        except Exception as exc:
-            logger.warning("Failed to load curriculum lesson %s: %s", lesson_code, exc)
-
-    _CURRICULUM_TITLE_CACHE = result
-    logger.info("Loaded %d curriculum titles for fuzzy-match fallback", len(result))
-    return result
-
-
-def _normalize_title(s: str) -> str:
-    """Normalize punctuation for fuzzy matching.
-
-    Covers two known mismatch classes:
-    1. Half-width ASCII vs full-width CJK (e.g. '?' vs '？')
-    2. Dash variants: YAML uses '──' (BOX DRAWINGS LIGHT HORIZONTAL U+2500×2)
-       while Gemini OCR returns '——' (EM DASH U+2014×2). Collapse both to '—'.
-    """
-    return (
-        s.replace("?", "？")
-         .replace("!", "！")
-         .replace(",", "，")
-         .replace("──", "—")   # BOX DRAWINGS ×2 → em-dash
-         .replace("——", "—")   # EM DASH ×2 → em-dash
-         .replace("─", "—")    # single BOX DRAWINGS → em-dash
-         .strip()
-    )
-
-
-def _fuzzy_match_title(extracted_title: str) -> list[LessonCandidate]:
-    """Match extracted_title against all known curriculum titles using difflib.
-
-    Args:
-        extracted_title: Title string extracted by Gemini OCR.
-
-    Returns:
-        Up to 1 LessonCandidate:
-        - ratio >= 0.85 → confidence 0.95
-        - ratio 0.70-0.84 → confidence 0.70
-        - ratio < 0.70 → empty list
-    """
-    if not extracted_title:
-        return []
-
-    curriculum = _load_curriculum_titles()
-    if not curriculum:
-        return []
-
-    best_code = ""
-    best_ratio = 0.0
-    normalized_extracted = _normalize_title(extracted_title)
-
-    for code, info in curriculum.items():
-        ratio = difflib.SequenceMatcher(
-            None, normalized_extracted, _normalize_title(info["title"])
-        ).ratio()
-        if ratio > best_ratio:
-            best_ratio = ratio
-            best_code = code
-
-    if best_ratio < 0.70:
-        logger.info(
-            "Fuzzy match: extracted_title=%r best_ratio=%.3f < 0.70 — no match",
-            extracted_title,
-            best_ratio,
-        )
-        return []
-
-    info = curriculum[best_code]
-    confidence = 0.95 if best_ratio >= 0.85 else 0.70
-
-    # lesson_id is the numeric part of the code (e.g. G5-L25 → 25)
-    try:
-        lesson_id = int(best_code.split("-L")[1])
-    except (IndexError, ValueError):
-        lesson_id = 0
-
-    logger.info(
-        "Fuzzy match: extracted_title=%r → %s %r ratio=%.3f conf=%.2f",
-        extracted_title,
-        best_code,
-        info["title"],
-        best_ratio,
-        confidence,
-    )
-    resolved_grade_code = info["grade_code"]
-    return [
-        LessonCandidate(
-            lesson_id=lesson_id,
-            grade_code=resolved_grade_code,
-            title=info["title"],
-            confidence=confidence,
-            reasoning=f"標題模糊匹配 ratio={best_ratio:.2f}",
-            story_id=_resolve_story_id(resolved_grade_code),
-        )
-    ]
-
-
-def _build_identification_prompt(lessons: list[dict]) -> str:
-    """Build the structured Gemini prompt for lesson identification.
-
-    With the full 158-lesson corpus, we list only grade_code + title per entry
-    (no story snippet or vocab) to keep prompt size manageable (~1300 tokens).
-    Gemini matches by reading the visible title from the worksheet image.
-    """
-    lesson_list = "\n".join(
-        f"  - {l['grade_code']}: \"{l['title']}\""
-        for l in lessons
-    )
-    n = len(lessons)
-
-    return f"""You are an AI reading assistant for a Taiwanese elementary/junior-high school platform.
-
-A student uploaded a photo of a paper worksheet. Your task is to identify which of the following {n} known lessons this worksheet belongs to, by reading visible text in the image.
-
-Known lessons:
-{lesson_list}
-
-Instructions:
-1. Read all visible text in the image, especially the large title text at the top.
-2. Extract the exact title you can read from the image. Put it in "extracted_title".
-3. Find the best matching lesson from the known list by comparing the extracted title verbatim.
-   IMPORTANT: If the extracted_title matches any known lesson title exactly or near-exactly,
-   that candidate MUST have confidence >= 0.9. Do NOT assign low confidence to an exact match.
-4. Return AT MOST 3 best-matching candidates, sorted by confidence (highest first).
-5. Use confidence 0.0–1.0 where:
-   - 0.95+ = extracted_title matches the lesson title exactly (character-for-character)
-   - 0.85–0.94 = extracted_title is nearly identical (1-2 char difference)
-   - 0.7–0.84 = partial title match
-   - 0.4–0.69 = content/theme match without title
-   - <0.4 = weak or speculative
-6. Keep each `reasoning` to ≤ 15 Chinese characters. Be terse.
-
-Respond ONLY with valid JSON in this exact format (no markdown, no explanation outside JSON):
-{{
-  "extracted_title": "<title text you could read from the image, or empty string>",
-  "candidates": [
-    {{
-      "lesson_id": <integer from grade_code suffix, e.g. G5-L25 → 25>,
-      "grade_code": "<string e.g. G5-L25>",
-      "title": "<string matching the known lesson title exactly>",
-      "confidence": <float 0.0-1.0>,
-      "reasoning": "<≤15 Chinese characters>"
-    }}
-  ]
-}}
-
-If the image is too blurry or no text is readable, return:
-{{
-  "extracted_title": "",
-  "candidates": [],
-  "error": "image_unclear"
-}}"""
 
 
 async def identify_lesson_from_image(image_bytes: bytes, mime_type: str = "image/jpeg") -> list[LessonCandidate]:
@@ -553,6 +256,10 @@ def identify_lesson_from_hint(lesson_code: str) -> list[LessonCandidate]:
     Returns:
         A single LessonCandidate with confidence=1.0, or [] if the code is not
         found in the curriculum catalog.
+
+    Note: Defined here (not delegated to omo_lesson_catalog) so that existing
+    tests can monkeypatch _load_curriculum_titles and _resolve_story_id on this
+    module (#1886 backward-compat).
     """
     curriculum = _load_curriculum_titles()
     if not curriculum:
